Finalized avatar loading.
So the process: something calls avatarImageForKey with a block that retrns the image. The method first checks cache, then the Documents folder, and if that find anything then it will fetch one from robohash.org. If *that* fails then it supplies the default avatar (which I updated to make square, and 48x48 (with @2x).

I love how well this turned out. It's beautiful.