Redid local notifcations
TWMessageBar is still used for in-app messages.

UILocalNotifications work for when we're outside the app.

Badge numbers work, but opening app from a local notification doesn't go to a specific window yet.

Still need to set default notifcation style to banners, and user can set it to alert views through Settings later.